more class practice

+++ class_prac/circle_math.py
@@ -8,7 +8,7 @@ class Circle():
     def perimeter(self):
         return round(2*self.radius*3.14, 3)
 
-x = int(input("Enter the radius of a circle: "))
+x = float(input("Enter the radius of a circle: "))
 NewCircle = Circle(x)
 print(NewCircle.area())
 print(NewCircle.perimeter())
